Module moteur_jeu_video::prelude

The elements needed to create a game

Macros

im_str

Structs

DevicesState

EngineError

EventLoopProxy
Used to send custom events to EventLoop.

Game
The Game structure It owns everything. When it created, it owns the EventLoop, but that one gets moved out uppon runing.

GameState
A GameState contains a SPECS world, and can be stacked and popped out of the GameStateStack GameStates, in the game, have a logic to be run and their rendering to be done.

Params
Owns the rendering parameters.

Scene
A scene contains pointers to existing ressources and datas to place them in the space.

Similarity
The matrix of a similarity. Used to displace, rotate and resize a 3D object. https://en.wikipedia.org/wiki/Similarity_(geometry)

Ui
A temporary reference for building the user interface for one frame

Window
Builder for a window

Enums

Condition
Condition for applying a setting

GameEvent

Key

Light
Different kinds of Light (directional and non directional)

LogicBehavior
Defines the logic behavior of the GameState

RenderBehavior
Defines the render behavior of the GameState

Functions

get_ressources_path
Finds the path of the ressources folder

rotation
A rotation 4 * 4 matrix created from an axis of 3 scalars and an angle expressed in radians.

translation
A 4 * 4 translation matrix created from the scaling factor on each axis.

vec3
Creates a new 3D vector.

vec4
Creates a new 4D vector.

Type Definitions

TMat4
A 4x4 matrix with components of type N.

Vector3
A stack-allocated, 3-dimensional column vector.